+++ diesel/src/sqlite/mod.rs
@@ -25,6 +25,8 @@ pub use self::connection::SerializedDatabase;
 pub use self::connection::SqliteBindValue;
 pub use self::connection::SqliteConnection;
 pub use self::connection::SqliteLimit;
+pub use self::connection::SqliteTraceEvent;
+pub use self::connection::SqliteTraceFlags;
 pub use self::connection::SqliteValue;
 pub use self::connection::authorizer;
 pub use self::connection::sqlite_blob::SqliteReadOnlyBlob;